+++ @xen-orchestra/lite/src/components/vm/VmsActionsBar.vue
@@ -12,15 +12,13 @@
     <MenuItem :icon="faPowerOff">
       {{ $t("change-power-state") }}
       <template #submenu>
-        <VmsPowerActionsMenu :vm-refs="selectedRefs" />
+        <VmActionPowerStateItems :vm-refs="selectedRefs" />
       </template>
     </MenuItem>
     <MenuItem v-tooltip="$t('coming-soon')" :icon="faRoute">
       {{ $t("migrate") }}
     </MenuItem>
-    <MenuItem v-tooltip="$t('coming-soon')" :icon="faCopy">
-      {{ $t("copy") }}
-    </MenuItem>
+    <VmActionCopyItem :selected-refs="selectedRefs" />
     <MenuItem v-tooltip="$t('coming-soon')" :icon="faEdit">
       {{ $t("edit-config") }}
     </MenuItem>
@@ -60,13 +58,13 @@
 import AppMenu from "@/components/menu/AppMenu.vue";
 import MenuItem from "@/components/menu/MenuItem.vue";
 import UiButton from "@/components/ui/UiButton.vue";
+import VmActionCopyItem from "@/components/vm/VmActionItems/VmActionCopyItem.vue";
 import { useUiStore } from "@/stores/ui.store";
-import VmsPowerActionsMenu from "@/components/vm/VmsPowerActionsMenu.vue";
+import VmActionPowerStateItems from "@/components/vm/VmActionItems/VmActionPowerStateItems.vue";
 import { vTooltip } from "@/directives/tooltip.directive";
 import {
   faCamera,
   faCode,
-  faCopy,
   faDisplay,
   faEdit,
   faEllipsis,